requirements.txt
Django>=2.2.12,<2.3
djangorestframework>=3.11.0,<3.12.0